term extraction in separate file
