+++ nafigator/__init__.py
@@ -6,6 +6,7 @@ from .cli import *
 from .const import *
 from .utils import *
 from .nafdocument import *
+from .term_extraction import *
 
 from .parse2naf import *
 from .parse2folia import *
